leds:qpnp-flash-v2: Add support for flash realtime control
In the default design, every time we change the brightness
in torch node it gets reflected only when the enable bit
in the switch node is toggled from 0 to 1. Add new DT property
to optionally allow brightness change to reflect in realtime,
for as long as the switch node enable bit remains set.

Signed-off-by: Kavya Nunna <quic_knunna@quicinc.com>
Change-Id: If4cd4dbea611ab6c13ec8ac2ec16d6b39fed1e20
Signed-off-by: TogoFire <togofire@mailfence.com>
Signed-off-by: Edwiin Kusuma Jaya <kutemeikito0905@gmail.com>
Signed-off-by: neophyte <neophyteplayground@gmail.com>
Signed-off-by: neophyte404 <260553983+neophyte404@users.noreply.github.com>